gradio/requirements.txt
huggingface-hub==0.28.1
GitPython==3.1.43
gradio==5.23.3
pdf2image==1.17.0
dill==0.4.0
